fixed server side render issue caused by TG SDK
